+++ R/pype_from_seurat.R
@@ -5,6 +5,7 @@
 
 
 
+
 #' Convert Seurat to cellpypes object. 
 #' 
 #' Start cellpyping a Seurat object.
@@ -13,28 +14,47 @@
 #'   
 #'
 #' @param seurat A Seurat object.
+#' @param graph_name Supply one of the graphs. To see options, type 
+#' `names(seurat@graphs)`. If left empty (`NULL`, the default),
+#' `pype_from_seurat` will try to guess the correct name for you.
 #'
 #' @return A cellpypes object.
 #' @template cellpypes_obj
 #' @export
-pype_from_seurat <- function(seurat) {
+pype_from_seurat <- function(seurat, graph_name=NULL) {
   seurat_status <- requireNamespace("Seurat", quietly=TRUE) &&
     requireNamespace("SeuratObject", quietly = TRUE)
   if(!seurat_status) stop("Install Seurat to use this function.")
   stopifnot(inherits(seurat, "Seurat"))
+  graph_name_status = rlang::is_scalar_character(graph_name) || is.null(graph_name)
+  stopifnot("graph_name must be a character scalar."=graph_name_status)
   
-  
-  # I pick graphs according to this wish list (order matters):
-  # WNN_snn, SCT_snn, RNA_snn.
-  all_graphs <- names(seurat@graphs)
-  if (is.null(all_graphs)) {
-    seurat <- Seurat::FindNeighbors(seurat)
+  if(is.null(graph_name)) {
+    # I pick graphs according to this wish list (order matters):
+    # WNN_snn, SCT_snn, RNA_snn.
     all_graphs <- names(seurat@graphs)
+    if (is.null(all_graphs)) {
+      seurat <- Seurat::FindNeighbors(seurat)
+      all_graphs <- names(seurat@graphs)
+    }
+    
+    # order matters:
+    seurat_graph_prefixes <- c("integrated","WNN", "SCT", "RNA") 
+    graph_choices <- c(paste0(seurat_graph_prefixes, "_snn"),
+                       paste0(seurat_graph_prefixes, "_nn"))
+    
+    graph_choices <- graph_choices[graph_choices %in% names(seurat@graphs)]
+    if(length(graph_choices)>0) {
+      graph_choice <- graph_choices[1]
+    } else {
+      stop("No neighbor graph found. Try passing names(seurat@graphs) to argument graph_name.")
+    }
+    
+  } else { # user-supplied graph_name
+    stopifnot("graph_name not present in names(seurat@graphs)"=graph_name %in% names(seurat@graphs))
+    graph_choice <- graph_name
   }
-  snn_graphs <- all_graphs[grepl("_snn$", all_graphs)]
-  graph_choices <- paste(c("WNN", "SCT", "RNA"), "snn", sep="_")
-  graph_choice <- graph_choices[graph_choices %in% snn_graphs][1]
-  
+ 
   # pick embedding according to this wish list order: umap, then tsne
   if ("umap" %in% names(seurat@reductions)) {
     dimension_names <- c("UMAP_1", "UMAP_2")